Add files via upload
Chunks of code I would use for various aggregation of data that leadership would regularly request. Data was frequently extracted using the REDCap API, but to protect the confidentiality of API tokens, it is shown here as being read from a csv format off the local device in a generic file path.